+++ lib/msal-angular/src/msal.module.ts
@@ -1,15 +1,12 @@
 import { Injectable, ModuleWithProviders, NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
-import { MsalService, MSAL_CONFIG, MSAL_CONFIG_ANGULAR } from "./msal.service";
+import { MsalService } from "./msal.service";
 import { MsalGuard } from "./msal-guard.service";
 import { BroadcastService } from "./broadcast.service";
 import { Configuration } from "msal";
 import { MsalAngularConfiguration, defaultMsalAngularConfiguration } from "./msal-angular.configuration";
+import { MSAL_CONFIG, MSAL_CONFIG_ANGULAR } from "./constants";
 
-Injectable()
-export class WindowWrapper extends Window {
-
-}
 @NgModule({
   imports: [CommonModule],
     declarations: [
@@ -33,11 +30,7 @@ export class MsalModule {
             provide: MSAL_CONFIG_ANGULAR,
             useValue: angularConfig
         },
-        MsalService,
-        {
-            provide: WindowWrapper,
-            useValue: window
-        }
+        MsalService
       ]
     }
   }