+++ lib/msal-node/test/crypto/PkceGenerator.spec.ts
@@ -2,7 +2,6 @@ import { PkceCodes } from '@azure/msal-common';
 import { PkceGenerator } from './../../src/crypto/PkceGenerator';
 
 describe('PkceGenerator', () => {
-
     const NUM_TESTS = 100;
     // tests generatePkceCodes
     test('generatePkceCodes() generates valid PKCE codes', async () => {